requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.1
PyQt5==5.15.9
pyinstaller==5.13.0